refactor(core): Centralize context management logic into src/context
