+++ packages/core/src/context/chatCompressionService.test.ts
